+++ CHANGELOG.md
@@ -6,9 +6,19 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+- Expand `currency` feature. See the [`currency` configuration].
+
 ### Added
 - `!promo` now uses hangout detection determined by the `irc/idle-detection/threshold` setting.
 - `!afterstream` command without argument now prints a help message ([#26]).
+- Added `gtav` module to interface with [ChaosMod]. See [`gtav` configuration].
+
+[`currency` configuration]: README.md#currency
+[`gtav` configuration]: README.md#gtav
+[ChaosMod]: https://github.com/udoprog/ChaosMod
+
+[#26]: https://github.com/udoprog/setmod/issues/26
 
 ### Changed
 - Fixed bug where `Settings` frontend would make the value into a string before sending it to backend.